memberportal/requirements.txt
django~=3.2.13
requests~=2.27.1
stripe==2.74.0
humanize~=4.1.0
django-constance~=2.9.0
django-picklefield~=3.0.1
django-cors-headers~=3.12.0
black==22.3.0
pre-commit==2.19.0
djangorestframework~=3.13.1
djangorestframework-simplejwt>=4.7.2
django-csp~=3.7
sentry-sdk==1.5.12
mysqlclient
git+https://github.com/mailchimp/mailchimp-marketing-python.git
channels~=3.0.4
zeroconf~=0.38.6
postmarker~=1.0
channels_redis~=3.4.0
ics~=0.7
python_http_client~=3.3.7
pwned-passwords-django~=1.6.0
celery~=5.2.7
django-celery-results~=2.3.1
django-celery-beat~=2.2.1
redis~=4.3.1
twilio~=7.9.2
icalevents~=0.1.25
